Add @Event annotations on event DTO classes combine KafkaConfigurations
