+++ bill/src/main/java/com/tungstun/bill/application/bill/event/BillPayed.java
@@ -1,5 +1,8 @@
 package com.tungstun.bill.application.bill.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record BillPayed(
         Long id,
         Long barId,
