+++ bill/src/main/java/com/tungstun/bill/port/messaging/config/KafkaConfig.java
@@ -1,6 +1,9 @@
-package com.tungstun.security.port.messaging.config;
+package com.tungstun.bill.port.messaging.config;
+
 
 import com.tungstun.common.messaging.KafkaConfigBase;
+import com.tungstun.common.messaging.KafkaMessageProducer;
+import org.apache.kafka.clients.admin.NewTopic;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
@@ -8,18 +11,26 @@ import org.springframework.kafka.config.KafkaListenerContainerFactory;
 import org.springframework.kafka.listener.CommonLoggingErrorHandler;
 import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
-import java.util.Set;
-
 @Configuration
-public class KafkaConsumerConfig extends KafkaConfigBase {
-    private static final Set<Class<?>> CLASSES = Set.of();
+public class KafkaConfig extends KafkaConfigBase {
+    private static final String TOPIC = "bill";
+
+    @Bean
+    public NewTopic bill() {
+        return new NewTopic(TOPIC, 1, (short) 1);
+    }
+
+    @Bean
+    public KafkaMessageProducer kafkaMessageProducer() {
+        return new KafkaMessageProducer(TOPIC);
+    }
 
     @Bean
     public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, String>>
     kafkaListenerContainerFactory() {
         ConcurrentKafkaListenerContainerFactory<String, String> factory = new ConcurrentKafkaListenerContainerFactory<>();
         factory.setCommonErrorHandler(new CommonLoggingErrorHandler());
-        factory.setConsumerFactory(defaultConsumerFactory(CLASSES));
+        factory.setConsumerFactory(defaultConsumerFactory());
         return factory;
     }
 }
