+++ bill/src/main/java/com/tungstun/bill/port/messaging/in/person/message/PersonDeleted.java
@@ -1,5 +1,8 @@
 package com.tungstun.bill.port.messaging.in.person.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record PersonDeleted(
         Long id) {
 }
