+++ bill/src/main/java/com/tungstun/bill/port/messaging/in/product/message/ProductDeleted.java
@@ -1,5 +1,8 @@
 package com.tungstun.bill.port.messaging.in.product.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record ProductDeleted(
         Long id) {
 }
