+++ bill/src/main/java/com/tungstun/bill/port/messaging/in/product/message/ProductUpdated.java
@@ -1,5 +1,8 @@
 package com.tungstun.bill.port.messaging.in.product.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record ProductUpdated(
         Long id,
         String name,
