+++ common/src/main/java/com/tungstun/common/messaging/KafkaMessageProducer.java
@@ -2,6 +2,7 @@ package com.tungstun.common.messaging;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.core.KafkaTemplate;
 
 
@@ -9,16 +10,18 @@ public class KafkaMessageProducer {
     private static final Logger LOG = LoggerFactory.getLogger(KafkaMessageProducer.class);
 
     private final String topic;
-    private final KafkaTemplate<String, Object> template;
 
-    public KafkaMessageProducer(String topic, KafkaTemplate<String, Object> template) {
+    @Autowired
+    private KafkaTemplate<String, Object> template;
+
+    public KafkaMessageProducer(String topic) {
         this.topic = topic;
-        this.template = template;
     }
 
     public void publish(Number id, Object data) {
         publish(String.valueOf(id), data);
     }
+
     public void publish(String id, Object data) {
         template.send(topic, id, data)
                 .addCallback(
@@ -27,5 +30,4 @@ public class KafkaMessageProducer {
                                 id, data, failure.getLocalizedMessage())
                 );
     }
-
 }
