+++ core/src/main/java/com/tungstun/core/application/bar/event/BarDeleted.java
@@ -1,4 +1,7 @@
 package com.tungstun.core.application.bar.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record BarDeleted(Long id) {
 }
