+++ core/src/main/java/com/tungstun/core/application/session/event/SessionDeleted.java
@@ -1,4 +1,7 @@
 package com.tungstun.core.application.session.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record SessionDeleted(Long id, Long barId) {
 }
