+++ core/src/main/java/com/tungstun/core/application/session/event/SessionEnded.java
@@ -1,7 +1,10 @@
 package com.tungstun.core.application.session.event;
 
+import com.tungstun.common.messaging.Event;
+
 import java.time.LocalDateTime;
 
+@Event
 public record SessionEnded(
         Long id,
         LocalDateTime endDate) {
