+++ core/src/main/java/com/tungstun/core/application/session/event/SessionLocked.java
@@ -1,5 +1,8 @@
 package com.tungstun.core.application.session.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record SessionLocked(
         Long id,
         Boolean isLocked) {
