+++ core/src/main/java/com/tungstun/core/port/messaging/config/KafkaConfig.java
@@ -1,8 +1,8 @@
-package com.tungstun.person.port.messaging.config;
-
+package com.tungstun.core.port.messaging.config;
 
 import com.tungstun.common.messaging.KafkaConfigBase;
-import com.tungstun.person.port.messaging.in.user.message.UserCreated;
+import com.tungstun.common.messaging.KafkaMessageProducer;
+import org.apache.kafka.clients.admin.NewTopic;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
@@ -10,20 +10,26 @@ import org.springframework.kafka.config.KafkaListenerContainerFactory;
 import org.springframework.kafka.listener.CommonLoggingErrorHandler;
 import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
-import java.util.Set;
-
 @Configuration
-public class KafkaConsumerConfig extends KafkaConfigBase {
-    private static final Set<Class<?>> CLASSES = Set.of(
-            UserCreated.class
-    );
+public class KafkaConfig extends KafkaConfigBase {
+    private static final String TOPIC = "core";
+
+    @Bean
+    public NewTopic core() {
+        return new NewTopic(TOPIC, 1, (short) 1);
+    }
+
+    @Bean
+    public KafkaMessageProducer kafkaMessageProducer() {
+        return new KafkaMessageProducer(TOPIC);
+    }
 
     @Bean
     public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, String>>
     kafkaListenerContainerFactory() {
         ConcurrentKafkaListenerContainerFactory<String, String> factory = new ConcurrentKafkaListenerContainerFactory<>();
         factory.setCommonErrorHandler(new CommonLoggingErrorHandler());
-        factory.setConsumerFactory(defaultConsumerFactory(CLASSES));
+        factory.setConsumerFactory(defaultConsumerFactory());
         return factory;
     }
 }
