+++ core/src/main/java/com/tungstun/core/port/messaging/in/bill/message/BillCreated.java
@@ -1,5 +1,8 @@
 package com.tungstun.core.port.messaging.in.bill.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record BillCreated(
         Long id,
         Long barId,
