--- core/src/main/java/com/tungstun/core/port/messaging/in/security/messages/UserCreated.java
@@ -1,5 +1,0 @@
-package com.tungstun.core.port.messaging.in.security.messages;
-
-public record UserCreated(Long id,
-                          String username) {
-}
