+++ person/src/main/java/com/tungstun/person/application/person/event/PersonDeleted.java
@@ -1,4 +1,7 @@
 package com.tungstun.person.application.person.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record PersonDeleted(Long personId) {
 }
