+++ person/src/main/java/com/tungstun/person/application/person/event/PersonUpdated.java
@@ -1,5 +1,8 @@
 package com.tungstun.person.application.person.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record PersonUpdated(
         Long personId,
         Long barId,
