+++ person/src/main/java/com/tungstun/person/port/messaging/config/KafkaConfig.java
@@ -1,0 +1,36 @@
+package com.tungstun.person.port.messaging.config;
+
+
+import com.tungstun.common.messaging.KafkaConfigBase;
+import com.tungstun.common.messaging.KafkaMessageProducer;
+import org.apache.kafka.clients.admin.NewTopic;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
+import org.springframework.kafka.config.KafkaListenerContainerFactory;
+import org.springframework.kafka.listener.CommonLoggingErrorHandler;
+import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
+
+@Configuration
+public class KafkaConfig extends KafkaConfigBase {
+    private static final String TOPIC = "person";
+
+    @Bean
+    public NewTopic person() {
+        return new NewTopic(TOPIC, 1, (short) 1);
+    }
+
+    @Bean
+    public KafkaMessageProducer kafkaMessageProducer() {
+        return new KafkaMessageProducer(TOPIC);
+    }
+
+    @Bean
+    public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, String>>
+    kafkaListenerContainerFactory() {
+        ConcurrentKafkaListenerContainerFactory<String, String> factory = new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setCommonErrorHandler(new CommonLoggingErrorHandler());
+        factory.setConsumerFactory(defaultConsumerFactory());
+        return factory;
+    }
+}
