+++ person/src/main/java/com/tungstun/person/port/messaging/in/user/message/UserCreated.java
@@ -1,5 +1,8 @@
 package com.tungstun.person.port.messaging.in.user.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record UserCreated(
         Long id,
         String username,
