+++ product/src/main/java/com/tungstun/product/application/product/event/ProductDeleted.java
@@ -1,5 +1,8 @@
 package com.tungstun.product.application.product.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record ProductDeleted(
         Long id,
         Long barId) {
