+++ security/src/main/java/com/tungstun/security/application/user/event/UserCreated.java
@@ -1,5 +1,8 @@
 package com.tungstun.security.application.user.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record UserCreated(
         Long id,
         String username,
