+++ security/src/main/java/com/tungstun/security/application/user/event/UserUpdated.java
@@ -1,5 +1,8 @@
 package com.tungstun.security.application.user.event;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record UserUpdated(
         Long id,
         String username,
