+++ security/src/main/java/com/tungstun/security/port/messaging/config/KafkaConfig.java
@@ -1,9 +1,7 @@
-package com.tungstun.core.port.messaging.config;
+package com.tungstun.security.port.messaging.config;
 
 import com.tungstun.common.messaging.KafkaConfigBase;
-import com.tungstun.core.port.messaging.in.bill.message.BillCreated;
-import com.tungstun.core.port.messaging.in.bill.message.BillDeleted;
-import com.tungstun.core.port.messaging.in.security.messages.UserCreated;
+import com.tungstun.common.messaging.KafkaMessageProducer;
 import org.apache.kafka.clients.admin.NewTopic;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,20 +10,18 @@ import org.springframework.kafka.config.KafkaListenerContainerFactory;
 import org.springframework.kafka.listener.CommonLoggingErrorHandler;
 import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
-import java.util.Set;
-
 @Configuration
-public class KafkaConsumerConfig extends KafkaConfigBase {
-    private static final String CONSUMING_TOPIC = "bill";
-    private static final Set<Class<?>> CLASSES = Set.of(
-            BillCreated.class,
-            BillDeleted.class,
-            UserCreated.class
-    );
+public class KafkaConfig extends KafkaConfigBase {
+    private static final String TOPIC = "security";
 
     @Bean
-    public NewTopic createNewTopic() {
-        return new NewTopic(CONSUMING_TOPIC, 1, (short) 1);
+    public NewTopic security() {
+        return new NewTopic(TOPIC, 1, (short) 1);
+    }
+
+    @Bean
+    public KafkaMessageProducer kafkaMessageProducer() {
+        return new KafkaMessageProducer(TOPIC);
     }
 
     @Bean
@@ -33,7 +29,7 @@ public class KafkaConsumerConfig extends KafkaConfigBase {
     kafkaListenerContainerFactory() {
         ConcurrentKafkaListenerContainerFactory<String, String> factory = new ConcurrentKafkaListenerContainerFactory<>();
         factory.setCommonErrorHandler(new CommonLoggingErrorHandler());
-        factory.setConsumerFactory(defaultConsumerFactory(CLASSES));
+        factory.setConsumerFactory(defaultConsumerFactory());
         return factory;
     }
 }
