--- security/src/main/java/com/tungstun/security/port/messaging/config/KafkaProducerConfig.java
@@ -1,30 +1,0 @@
-package com.tungstun.security.port.messaging.config;
-
-import com.tungstun.common.messaging.KafkaConfigBase;
-import com.tungstun.common.messaging.KafkaMessageProducer;
-import com.tungstun.security.application.user.event.UserCreated;
-import com.tungstun.security.application.user.event.UserUpdated;
-import org.apache.kafka.clients.admin.NewTopic;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.Set;
-
-@Configuration
-public class KafkaProducerConfig extends KafkaConfigBase {
-    private static final String TOPIC = "security";
-    private static final Set<Class<?>> CLASSES = Set.of(
-            UserCreated.class,
-            UserUpdated.class
-    );
-
-    @Bean
-    public NewTopic security() {
-        return new NewTopic(TOPIC, 1, (short) 1);
-    }
-
-    @Bean
-    public KafkaMessageProducer kafkaMessageProducer() {
-        return new KafkaMessageProducer(TOPIC, defaultKafkaTemplate(CLASSES));
-    }
-}
