+++ security/src/main/java/com/tungstun/security/port/messaging/in/core/message/BarCreated.java
@@ -1,5 +1,8 @@
 package com.tungstun.security.port.messaging.in.core.message;
 
+import com.tungstun.common.messaging.Event;
+
+@Event
 public record BarCreated(
         Long barId,
         Long userId) {
